Change column type

+++ tools/DatabaseSetup/mssqlnative/table-schema.sql
@@ -19,15 +19,15 @@ DROP SEQUENCE IF EXISTS unittest_seq;
 -- Testtable_1 is used to test the basic functionality of the meta functions
 -- It has a variety of data types but contains no data
 CREATE TABLE testtable_1 (
-	id INT IDENTITY(1,1) PRIMARY KEY,
+	id BIGINT IDENTITY(1,1) PRIMARY KEY,
 	varchar_field VARCHAR(20),
 	datetime_field DATETIME,
 	date_field DATE,
-	integer_field INT DEFAULT 0,
+	integer_field INTEGER DEFAULT 0,
 	decimal_field DECIMAL(12,2) DEFAULT 0,
 	boolean_field BIT DEFAULT 0,
 	empty_field VARCHAR(240) DEFAULT '',
-	number_run_field INT DEFAULT 0,
+	number_run_field BIGINT DEFAULT 0,
 );
 CREATE	unique index vdx1 ON testtable_1 (varchar_field);
 CREATE	unique index vdx2 ON testtable_1 (integer_field,date_field);
